chore(): removed the entity package
